Adds /api/update-emoji and vercel cron to call it daily
Adds a GET endpoint which updates the emoji when requested. Note that
this does not return the emoji, only the status of the request and (if
successful) the total number of emojis.

Also adds a simple vercel.json cron which will call this enpoint daily
to ensure the emoji are fresh. :)